+++ fifteen/package.json
@@ -4,19 +4,19 @@
   "private": true,
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.3",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
     "bootstrap-css-only": "^4.4.1",
     "lodash": "^4.17.21",
-    "mdbreact": "^5.0.1",
+    "mdbreact": "^5.0.2",
     "react": "^17.0.2",
     "react-awesome-modal": "^2.0.5",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.2",
-    "react-modal": "^3.12.1",
+    "react-modal": "^3.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "reactstrap": "^8.9.0",
