+++ fifteen/src/components/MyPageSide/MyPageSide.jsx
@@ -1,0 +1,21 @@
+import React, {Component} from 'react';
+import styles from './MyPage_side.module.css';
+import {Link, withRouter} from "react-router-dom";
+
+class MyPageSide extends Component {
+    render(){
+        return(
+            <div>
+                <div className={styles.MyPage_side}>
+                    <div className={styles.MyPage_side__category}>
+                        <div className={styles.MyPage_side__category__text}>주문내역조회</div>
+                        <Link to='/profile'><div className={styles.MyPage_side__category__text}>회원정보</div></Link>
+                        <Link to='/wishlist'><div className={styles.MyPage_side__category__text}>관심상품</div></Link>
+                        <Link to='/coupon'><div className={styles.MyPage_side__category__text}>쿠폰</div></Link>
+                    </div>
+                </div>
+            </div>
+        )
+    }
+}
+export default withRouter(MyPageSide);
