--- fifteen/src/components/login/login.module.css
@@ -1,105 +1,0 @@
-.login {
-    display: flex;
-    flex-direction: column;
-    width: 1000px;
-    align-items: flex-start;
-    height: 600px;
-    box-sizing: border-box;
-}
-
-.login__box {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 1000px;
-}
-
-.header__bar {
-    height: 1.5px;
-    width: 100%;
-    background-color: black;
-    margin-bottom: 10px;
-}
-
-.login__title {
-    font-size: 35px;
-    font-weight: bold;
-    margin-bottom: 100px;
-}
-
-.login__form__box {
-    display: flex;
-    flex-direction: column;
-    align-content: center;
-    width: 500px;
-}
-
-.login__form__content {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-}
-
-.id__box {
-    color: #8C8C8C;
-    width: 500px;
-    height: 60px;
-    background-color: white;
-    border-bottom : solid #8c8c8c 2px;
-    font-size: 20px;
-    box-sizing: border-box;
-    padding: 5px;
-    display: flex;
-    flex-direction: row;
-    justify-content: left;
-    margin-bottom: 10px;
-}
-
-.pw__box {
-    color: #8C8C8C;
-    width: 500px;
-    height: 60px;
-    background-color: white;
-    border-bottom : solid #8c8c8c 2px;
-    font-size: 20px;
-    box-sizing: border-box;
-    padding: 5px;
-    display: flex;
-    flex-direction: row;
-    justify-content: left;
-    margin-bottom: 10px;
-}
-
-.id__box .input {
-    border: none;
-    width: 92%;
-    height: 100%;
-    margin-left: 20px;
-    font-size: 20px;
-}
-
-.pw__box .input {
-    border: none;
-    width: 80%;
-    height: 100%;
-    margin-left: 20px;
-    font-size: 20px;
-}
-
-.join {
-    font-size: 17px;
-    margin-bottom: 25px;
-}
-
-.login__btn {
-    width: 500px;
-    height: 64px;
-    box-sizing: border-box;
-    padding: 10px;
-    text-align: center;
-    border-radius: 10px;
-    background-color: black;
-    color: white;
-    font-size: 24px;
-    font-weight: bold;
-}
