+++ fifteen/src/pages/coupon/couponPage.css
@@ -1,0 +1,105 @@
+.couponPage{
+    display: flex;
+    flex-direction: column;
+}
+.couponPage_line{
+    width: 1239px;
+    align-self: center;
+    height: 2px;
+    background: #282c34;
+}
+.couponPage_title{
+    display: flex;
+    justify-content: flex-start;
+    margin: 10px 0 90px 150px;
+    font-weight: bold;
+    font-size: 35px;
+}
+.couponPage_main{
+    display: flex;;
+    flex-direction: row;
+    margin-left: 50px;
+}
+
+.coupon{
+    display: flex;
+    flex-direction: column;
+    width: 1000px;
+}
+.couponTitle{
+    display: flex;
+    justify-self: flex-start;
+    font-weight: bold;
+    font-size: 20px;
+    margin-bottom: 15px;
+}
+.couponTheme{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    height: 53px;
+    align-items: center;
+    border-top: 1.5px solid #BBBBBB;
+    border-bottom: 1.5px solid #BBBBBB;
+}
+.couponThemeSeq{
+    width:150px;
+}
+.couponThemeName{
+    width:400px;
+}
+.couponThemePrice{
+    width: 150px;
+}
+.couponThemeTerm{
+    width:300px;
+}
+
+.couponBox{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    height: 55px;
+    align-items: center;
+    border-bottom: 1.5px solid #BBBBBB;
+}
+.couponBoxSeq{
+    width:150px;
+}
+.couponBoxName{
+    width:400px;
+}
+.couponBoxPrice{
+    width: 150px;
+}
+.couponBoxTerm{
+    width:300px;
+}
+
+.couponButton{
+    display: flex;
+    flex-direction: column;
+}
+
+.couponPaging{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    margin-top: 50px;
+}
+.couponPagingBefore{
+    height: 40px;
+    width: 40px;
+    border: 1px solid #DDDDDD;
+}
+.couponPagingThis{
+    height: 40px;
+    width: 40px;
+    border: 1px solid #DDDDDD;
+    font-size: 18px;
+}
+.couponPagingAfter{
+    height: 40px;
+    width: 40px;
+    border: 1px solid #DDDDDD;
+}
