--- fifteen/src/pages/coupon/coupon_page.css
@@ -1,22 +1,0 @@
-.couponPage{
-    display: flex;
-    flex-direction: column;
-}
-.couponPage_line{
-    width: 1239px;
-    align-self: center;
-    height: 2px;
-    background: #282c34;
-}
-.couponPage_title{
-    display: flex;
-    justify-content: flex-start;
-    margin: 10px 0 90px 150px;
-    font-weight: bold;
-    font-size: 35px;
-}
-.couponPage_main{
-    display: flex;;
-    flex-direction: row;
-    margin-left: 50px;
-}
