--- fifteen/src/pages/join/join_page.css
@@ -1,5 +1,0 @@
-.join_page {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
