--- fifteen/src/pages/login/login_page.module.css
@@ -1,5 +1,0 @@
-.login {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
