--- fifteen/src/pages/mypage/my_page.module.css
@@ -1,24 +1,0 @@
-.My_Page{
-    display: flex;
-    flex-direction: column;
-}
-.My_Page_line{
-    width: 1239px;
-    align-self: center;
-    height: 2px;
-    background: #282c34;
-}
-.My_page__title{
-    display: flex;
-    justify-content: flex-start;
-    margin: 10px 0 90px 150px;
-    font-weight: bold;
-    font-size: 35px;
-
-}
-
-.My_page__main{
-    display: flex;;
-    flex-direction: row;
-    margin-left: 50px;
-}
