+++ fifteen/src/pages/order/orderPage.css
@@ -1,0 +1,64 @@
+.orderPage{
+    display: flex;
+    flex-direction: column;
+}
+.orderPage__title{
+    display: flex;
+    justify-content: flex-start;
+    margin: 10px 0 90px 150px;
+    font-weight: bold;
+    font-size: 35px;
+
+}
+.orderPage__main{
+    display: flex;;
+    flex-direction: row;
+    margin-left: 50px;
+}
+.orderPage_line{
+    width: 1239px;
+    align-self: center;
+    height: 2px;
+    background: #282c34;
+}
+
+.order{
+    display: flex;
+    flex-direction: column;
+    width: 1000px;
+    height: 500px;
+}
+.order_info_title{
+    display: flex;
+    justify-content: flex-start;
+    font-weight: bold;
+    font-size: 20px;
+}
+.order_info_box{
+    margin-top: 15px;
+    height: 300px;
+    border-top: 1.5px solid #BBBBBB;
+    border-bottom: 1.5px solid #BBBBBB;
+    margin-bottom: 55px;
+}
+.order_paging{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+}
+.order_paging_before{
+    height: 40px;
+    width: 40px;
+    border: 1px solid #DDDDDD;
+}
+.order_paging_this{
+    height: 40px;
+    width: 40px;
+    border: 1px solid #DDDDDD;
+    font-size: 18px;
+}
+.order_paging_after{
+    height: 40px;
+    width: 40px;
+    border: 1px solid #DDDDDD;
+}
