--- fifteen/src/pages/order/order_page.css
@@ -1,23 +1,0 @@
-.orderPage{
-    display: flex;
-    flex-direction: column;
-}
-.orderPage__title{
-    display: flex;
-    justify-content: flex-start;
-    margin: 10px 0 90px 150px;
-    font-weight: bold;
-    font-size: 35px;
-
-}
-.orderPage__main{
-    display: flex;;
-    flex-direction: row;
-    margin-left: 50px;
-}
-.orderPage_line{
-    width: 1239px;
-    align-self: center;
-    height: 2px;
-    background: #282c34;
-}
