--- fifteen/src/pages/profile/profile_page.css
@@ -1,22 +1,0 @@
-.profilePage{
-    display: flex;
-    flex-direction: column;
-}
-.profilePage_line{
-    width: 1239px;
-    align-self: center;
-    height: 2px;
-    background: #282c34;
-}
-.profilePage_title{
-    display: flex;
-    justify-content: flex-start;
-    margin: 10px 0 90px 150px;
-    font-weight: bold;
-    font-size: 35px;
-}
-.profilePage_main{
-    display: flex;;
-    flex-direction: row;
-    margin-left: 50px;
-}
