--- fifteen/src/pages/wishlist/wishlist_page.css
@@ -1,22 +1,0 @@
-.wishlistPage{
-    display: flex;
-    flex-direction: column;
-}
-.wishlistPage_line{
-    width: 1239px;
-    align-self: center;
-    height: 2px;
-    background: #282c34;
-}
-.wishlistPage_title{
-    display: flex;
-    justify-content: flex-start;
-    margin: 10px 0 90px 150px;
-    font-weight: bold;
-    font-size: 35px;
-}
-.wishlistPage_main{
-    display: flex;;
-    flex-direction: row;
-    margin-left: 50px;
-}
